simplify the getFixedTiles method
We do not need to use reduce here. It's easier and simpler to read by just mutating the array in place as you loop through, and achieves the same goal of keeping track of changes to the array as you loop through. This improves performance as well by not copying the array each time.